+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [x]
+- Change login button
+
 ## [1.2.0]
 - Some improvments